refactor: 캠페인 리팩토링
