+++ src/main/java/be15fintomatokatchupbe/campaign/command/application/service/CampaignCommandService.java
@@ -10,7 +10,6 @@ import be15fintomatokatchupbe.client.command.application.support.ClientHelperSer
 import be15fintomatokatchupbe.client.command.domain.aggregate.ClientCompany;
 import be15fintomatokatchupbe.client.command.domain.aggregate.ClientManager;
 import be15fintomatokatchupbe.common.exception.BusinessException;
-import be15fintomatokatchupbe.file.domain.File;
 import be15fintomatokatchupbe.file.service.FileService;
 import be15fintomatokatchupbe.relation.service.HashInfCampService;
 import be15fintomatokatchupbe.relation.service.PipeInfClientManagerService;
@@ -21,10 +20,6 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
-import java.util.Optional;
 
 
 @AllArgsConstructor
@@ -34,7 +29,6 @@ public class CampaignCommandService {
     private final PipeUserService pipeUserService;
     private final PipeInfClientManagerService pipeInfClientManagerService;
     private final HashInfCampService hashInfCampService;
-    private final FileService fileService;
 
     private final ClientHelperService clientHelperService;
     private final CampaignHelperService campaignHelperService;
@@ -44,8 +38,6 @@ public class CampaignCommandService {
     private final CampaignStatusRepository campaignStatusRepository;
     private final PipelineRepository pipelineRepository;
     private final PipelineStepRepository pipelineStepRepository;
-    private final PipelineStatusRepository pipelineStatusRepository;
-    private final IdeaRepository ideaRepository;
 
     @Transactional
     public void createChance(Long userId, CreateChanceRequest request) {
@@ -104,241 +96,6 @@ public class CampaignCommandService {
 
         // 5. 해시태그 입력하기
         hashInfCampService.updateCampaignTags(campaign, request.getCategoryList());
-    }
-
-    @Transactional
-    public void createProposal(Long userId, CreateProposalRequest request) {
-        /* 외부 엔티티 가져오기
-         * : 광고 담당자, 캠페인, 파이프라인 단계 */
-
-        //. 광고 담당자 가져오기
-        ClientManager clientManager =
-                clientHelperService.findValidClientManager(request.getClientManagerId());
-
-        // 캠페인 가져오기
-        Campaign campaign =
-                campaignHelperService.findValidCampaign(request.getCampaignId());
-
-        // 파이프라인 단계 가져오기
-        PipelineStep pipelineStep =
-                pipelineStepRepository.findById(PipelineStepConstants.PROPOSAL)
-                        .orElseThrow(() -> new BusinessException(CampaignErrorCode.PIPELINE_STEP_NOT_FOUND));
-
-        // 파이프라인 상태 가져오기
-        PipelineStatus pipelineStatus =
-                pipelineStatusRepository.findById(request.getPipelineStatusId())
-                        .orElseThrow(() -> new BusinessException(CampaignErrorCode.PIPELINE_STATUS_NOT_FOUND));
-
-        // 작성자 가져오기
-        User writer = userHelperService.findValidUser(userId);
-
-        /* DB에 값 입력하기*/
-        // 파이프라인 생성 - 저장
-        Pipeline pipeline = Pipeline.builder()
-                .writer(writer)
-                .pipelineStep(pipelineStep)
-                .pipelineStatus(pipelineStatus)
-                .name(request.getName())
-                .requestAt(request.getRequestAt())
-                .startedAt(request.getStartedAt())
-                .endedAt(request.getEndedAt())
-                .presentedAt(request.getPresentedAt())
-                .campaign(campaign)
-                .content(request.getContent())
-                .notes(request.getNotes())
-                .build();
-
-        pipelineRepository.save(pipeline);
-
-        List<Idea> ideaList = Optional.ofNullable(request.getIdeaList())
-                .orElse(List.of()) // null인 경우 빈 리스트로 처리
-                .stream()
-                .map(idea -> Idea.builder()
-                        .content(idea.getContent())
-                        .user(writer)
-                        .pipeline(pipeline)
-                        .build())
-                .toList();
-
-        if (!ideaList.isEmpty()) {
-            ideaRepository.saveAll(ideaList);
-        }
-
-        /* 부가 데이터 각 테이블에 저장하기
-         * : 광고 담당자, 인플루언서 강점-비고, 담당자*/
-
-        /* 광고 담당자 */
-        pipeInfClientManagerService.saveClientManager(clientManager, pipeline);
-
-        /* 인플루언서 정보 */
-        pipeInfClientManagerService.saveInfluencerInfo(request.getInfluencerList(), pipeline);
-
-        /* 담당자*/
-        pipeUserService.saveUserList(request.getUserId(), pipeline);
-    }
-
-    @Transactional
-    public void createQuotation(Long userId, CreateQuotationRequest request) {
-        ClientManager clientManager = clientHelperService.findValidClientManager(request.getClientManagerId());
-        Campaign campaign = campaignHelperService.findValidCampaign(request.getCampaignId());
-        PipelineStep pipelineStep = pipelineStepRepository.findById(PipelineStepConstants.QUOTATION)
-                .orElseThrow(() -> new BusinessException(CampaignErrorCode.PIPELINE_STEP_NOT_FOUND));
-        PipelineStatus pipelineStatus = pipelineStatusRepository.findById(request.getPipelineStatusId())
-                .orElseThrow(() -> new BusinessException(CampaignErrorCode.PIPELINE_STATUS_NOT_FOUND));
-        User writer = userHelperService.findValidUser(userId);
-
-        Pipeline pipeline = Pipeline.builder()
-                .writer(writer)
-                .pipelineStep(pipelineStep)
-                .pipelineStatus(pipelineStatus)
-                .name(request.getName())
-                .requestAt(request.getRequestAt())
-                .startedAt(request.getStartedAt())
-                .endedAt(request.getEndedAt())
-                .presentedAt(request.getPresentedAt())
-                .campaign(campaign)
-                .content(request.getContent())
-                .notes(request.getNotes())
-                .expectedRevenue(request.getExpectedRevenue())
-                .expectedProfit(request.getExpectedProfit())
-                .availableQuantity(request.getAvailableQuantity())
-                .build();
-        pipelineRepository.save(pipeline);
-
-        List<Idea> ideaList = Optional.ofNullable(request.getIdeaList())
-                .orElse(List.of()) // null인 경우 빈 리스트로 처리
-                .stream()
-                .map(idea -> Idea.builder()
-                        .content(idea.getContent())
-                        .user(writer)
-                        .pipeline(pipeline)
-                        .build())
-                .toList();
-
-        if (!ideaList.isEmpty()) {
-            ideaRepository.saveAll(ideaList);
-        }
-
-        pipeInfClientManagerService.saveClientManager(clientManager, pipeline);
-        pipeInfClientManagerService.saveInfluencer(request.getInfluencerId(), pipeline);
-        pipeUserService.saveUserList(request.getUserId(), pipeline);
-    }
-
-    @Transactional
-    public void createContract(Long userId, CreateContractRequest request, List<MultipartFile> files) {
-        ClientManager clientManager = clientHelperService.findValidClientManager(request.getClientManagerId());
-        Campaign campaign = campaignHelperService.findValidCampaign(request.getCampaignId());
-        PipelineStep pipelineStep = pipelineStepRepository.findById(PipelineStepConstants.CONTRACT)
-                .orElseThrow(() -> new BusinessException(CampaignErrorCode.PIPELINE_STEP_NOT_FOUND));
-        PipelineStatus pipelineStatus = pipelineStatusRepository.findById(request.getPipelineStatusId())
-                .orElseThrow(() -> new BusinessException(CampaignErrorCode.PIPELINE_STATUS_NOT_FOUND));
-        User writer = userHelperService.findValidUser(userId);
-
-        Pipeline pipeline = Pipeline.builder()
-                .pipelineStep(pipelineStep)
-                .pipelineStatus(pipelineStatus)
-                .campaign(campaign)
-                .name(request.getName())
-                .requestAt(request.getRequestAt())
-                .startedAt(request.getStartedAt())
-                .endedAt(request.getEndedAt())
-                .presentedAt(request.getPresentedAt())
-                .content(request.getContent())
-                .notes(request.getNotes())
-                .writer(writer)
-                .expectedRevenue(request.getExpectedRevenue())
-                .expectedProfit(request.getExpectedProfit())
-                .availableQuantity(request.getAvailableQuantity())
-                .build();
-
-        pipelineRepository.save(pipeline);
-
-        List<Idea> ideaList = Optional.ofNullable(request.getIdeaList())
-                .orElse(List.of()) // null인 경우 빈 리스트로 처리
-                .stream()
-                .map(idea -> Idea.builder()
-                        .content(idea.getContent())
-                        .user(writer)
-                        .pipeline(pipeline)
-                        .build())
-                .toList();
-
-        if (!ideaList.isEmpty()) {
-            ideaRepository.saveAll(ideaList);
-        }
-
-        /* 파일 저장*/
-        if(files != null && !files.isEmpty()){
-            // 1. 파일 S3에 올리고 돌려 받기
-            List<File> fileList = fileService.uploadFile(files);
-            fileList.forEach(file -> file.setPipeline(pipeline));
-
-            // 2. 파일 DB에 저장하기
-            fileService.saveFile(fileList);
-        }
-
-        pipeInfClientManagerService.saveClientManager(clientManager, pipeline);
-        pipeInfClientManagerService.saveInfluencer(request.getInfluencerId(), pipeline);
-        pipeUserService.saveUserList(request.getUserId(), pipeline);
-    }
-
-    @Transactional
-    public void createRevenue(Long userId, CreateRevenueRequest request, List<MultipartFile> files) {
-        ClientManager clientManager = clientHelperService.findValidClientManager(request.getClientManagerId());
-        Campaign campaign = campaignHelperService.findValidCampaign(request.getCampaignId());
-        PipelineStep pipelineStep = pipelineStepRepository.findById(PipelineStepConstants.REVENUE)
-                .orElseThrow(() -> new BusinessException(CampaignErrorCode.PIPELINE_STEP_NOT_FOUND));
-        PipelineStatus pipelineStatus = pipelineStatusRepository.findById(request.getPipelineStatusId())
-                .orElseThrow(() -> new BusinessException(CampaignErrorCode.PIPELINE_STATUS_NOT_FOUND));
-        User writer = userHelperService.findValidUser(userId);
-
-        campaign.setProductPrice(request.getProductPrice());
-        campaignRepository.save(campaign);
-
-        Pipeline pipeline = Pipeline.builder()
-                .pipelineStep(pipelineStep)
-                .pipelineStatus(pipelineStatus)
-                .campaign(campaign)
-                .name(request.getName())
-                .requestAt(request.getRequestAt())
-                .startedAt(request.getStartedAt())
-                .endedAt(request.getEndedAt())
-                .presentedAt(request.getPresentedAt())
-                .content(request.getContent())
-                .notes(request.getNotes())
-                .writer(writer)
-                .build();
-
-        pipelineRepository.save(pipeline);
-
-        List<Idea> ideaList = Optional.ofNullable(request.getIdeaList())
-                .orElse(List.of()) // null인 경우 빈 리스트로 처리
-                .stream()
-                .map(idea -> Idea.builder()
-                        .content(idea.getContent())
-                        .user(writer)
-                        .pipeline(pipeline)
-                        .build())
-                .toList();
-
-        if (!ideaList.isEmpty()) {
-            ideaRepository.saveAll(ideaList);
-        }
-
-        /* 파일 저장*/
-        if(files != null && !files.isEmpty()){
-            // 1. 파일 S3에 올리고 돌려 받기
-            List<File> fileList = fileService.uploadFile(files);
-            fileList.forEach(file -> file.setPipeline(pipeline));
-
-            // 2. 파일 DB에 저장하기
-            fileService.saveFile(fileList);
-        }
-
-        pipeInfClientManagerService.saveClientManager(clientManager, pipeline);
-        pipeInfClientManagerService.saveInfluencerRevenue(request.getInfluencerList(), pipeline);
-        pipeUserService.saveUserList(request.getUserId(), pipeline);
-
     }
 
     // 캠페인 상세 수정
