+++ src/main/java/be15fintomatokatchupbe/campaign/command/application/service/ListupCommandService.java
@@ -1,0 +1,11 @@
+package be15fintomatokatchupbe.campaign.command.application.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class ListupCommandService {
+}
